requirements.txt
matplotlib
pandas 
streamlit
plotly.express 
numpy 
streamlit-bokeh-events